fix: add a common method to get meta from request
